Changed delay parameter for ent_fire
Originally, the delay param was an integer, and supplying any decimal value as a delay (e.g. 2.5 seconds) would simply round down. (2 seconds).

+++ sp/src/game/server/baseentity.cpp
@@ -5341,7 +5341,7 @@ public:
 		{
 			const char *target = "", *action = "Use";
 			variant_t value;
-			int delay = 0;
+			float delay = 0;
 
 			target = STRING( AllocPooledString(command.Arg( 1 ) ) );
